+++ hardhat.config.ts
@@ -8,14 +8,22 @@ import "@nomicfoundation/hardhat-foundry";
 dotenv.config();
 
 const config: HardhatUserConfig = {
-  solidity: "0.8.19",
+  solidity: {
+    compilers: [
+      {
+        version: "0.8.24",
+        settings: {
+          evmVersion: "paris"
+        },
+      }
+    ]
+  },
   paths: {
     sources: "./src",
   },
   docgen: {
     pages: "files",
     pageExtension: ".mdx",
-    templates: "custom_templates",
     exclude: ["test"],
     outputDir: "api_reference/solidity",
   },