+++ src/ccip/CCIPLocalSimulator.sol
@@ -6,8 +6,12 @@ import {LinkToken} from "../shared/LinkToken.sol";
 import {BurnMintERC677Helper} from "./BurnMintERC677Helper.sol";
 import {MockCCIPRouter} from "@chainlink/contracts-ccip/src/v0.8/ccip/test/mocks/MockRouter.sol";
 import {IRouterClient} from "@chainlink/contracts-ccip/src/v0.8/ccip/interfaces/IRouterClient.sol";
-import {IERC20} from "@chainlink/contracts-ccip/src/v0.8/vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/IERC20.sol";
-import {SafeERC20} from "@chainlink/contracts-ccip/src/v0.8/vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/utils/SafeERC20.sol";
+import {IERC20} from
+    "@chainlink/contracts-ccip/src/v0.8/vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/IERC20.sol";
+import {SafeERC20} from
+    "@chainlink/contracts-ccip/src/v0.8/vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/utils/SafeERC20.sol";
+import {IOwner} from "@chainlink/contracts-ccip/src/v0.8/ccip/interfaces/IOwner.sol";
+import {IGetCCIPAdmin} from "@chainlink/contracts-ccip/src/v0.8/ccip/interfaces/IGetCCIPAdmin.sol";
 
 /// @title CCIPLocalSimulator
 /// @notice This contract simulates local CCIP (Cross-Chain Interoperability Protocol) operations for testing and development purposes.
@@ -36,6 +40,8 @@ contract CCIPLocalSimulator {
     /// @notice The list of supported token addresses
     address[] internal s_supportedTokens;
 
+    error CCIPLocalSimulator__MsgSenderIsNotTokenOwner();
+
     /**
      * @notice Constructor to initialize the contract and pre-deployed token instances
      */
@@ -52,10 +58,29 @@ contract CCIPLocalSimulator {
 
     /**
      * @notice Allows user to support any new token, besides CCIP BnM and CCIP LnM, for cross-chain transfers.
+     *         Reverts if token does not implement owner() function.
+     *         Reverts if the caller is not the token owner.
      *
      * @param tokenAddress - The address of the token to add to the list of supported tokens.
      */
-    function supportNewToken(address tokenAddress) external {
+    function supportNewTokenViaOwner(address tokenAddress) external {
+        if (msg.sender != IOwner(tokenAddress).owner()) {
+            revert CCIPLocalSimulator__MsgSenderIsNotTokenOwner();
+        }
+        s_supportedTokens.push(tokenAddress);
+    }
+
+    /**
+     * @notice Allows user to support any new token, besides CCIP BnM and CCIP LnM, for cross-chain transfers.
+     *         Reverts if token does not implement getCCIPAdmin() function.
+     *         Reverts if the caller is not the token CCIPAdmin.
+     *
+     * @param tokenAddress - The address of the token to add to the list of supported tokens.
+     */
+    function supportNewTokenViaGetCCIPAdmin(address tokenAddress) external {
+        if (msg.sender != IGetCCIPAdmin(tokenAddress).getCCIPAdmin()) {
+            revert CCIPLocalSimulator__MsgSenderIsNotTokenOwner();
+        }
         s_supportedTokens.push(tokenAddress);
     }
 
@@ -66,9 +91,7 @@ contract CCIPLocalSimulator {
      *
      * @return supported - Returns true if `chainSelector` is supported by the simulator.
      */
-    function isChainSupported(
-        uint64 chainSelector
-    ) public pure returns (bool supported) {
+    function isChainSupported(uint64 chainSelector) public pure returns (bool supported) {
         supported = chainSelector == CHAIN_SELECTOR;
     }
 
@@ -79,9 +102,7 @@ contract CCIPLocalSimulator {
      *
      * @return tokens - Returns a list of token addresses that are supported for cross-chain transfers by the simulator.
      */
-    function getSupportedTokens(
-        uint64 chainSelector
-    ) external view returns (address[] memory tokens) {
+    function getSupportedTokens(uint64 chainSelector) external view returns (address[] memory tokens) {
         if (!isChainSupported(chainSelector)) {
             return new address[](0);
         }
@@ -97,10 +118,7 @@ contract CCIPLocalSimulator {
      *
      * @return success - Returns `true` if the transfer of tokens was successful, otherwise `false`.
      */
-    function requestLinkFromFaucet(
-        address to,
-        uint256 amount
-    ) external returns (bool success) {
+    function requestLinkFromFaucet(address to, uint256 amount) external returns (bool success) {
         success = i_linkToken.transfer(to, amount);
     }