+++ src/test/ccip/CCIPReceiver_Unsafe.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.19;
+pragma solidity 0.8.24;
 
 import {CCIPReceiver} from "@chainlink/contracts-ccip/src/v0.8/ccip/applications/CCIPReceiver.sol";
 import {Client} from "@chainlink/contracts-ccip/src/v0.8/ccip/libraries/Client.sol";
@@ -9,9 +9,7 @@ contract CCIPReceiver_Unsafe is CCIPReceiver {
 
     constructor(address router) CCIPReceiver(router) {}
 
-    function _ccipReceive(
-        Client.Any2EVMMessage memory message
-    ) internal override {
+    function _ccipReceive(Client.Any2EVMMessage memory message) internal override {
         text = abi.decode(message.data, (string));
     }
 }