+++ test/smoke/ccip/TokenTransferor.t.sol
@@ -1,9 +1,14 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.19;
+pragma solidity ^0.8.24;
 
 import {Test, console2} from "forge-std/Test.sol";
 import {TokenTransferor} from "../../../src/test/ccip/TokenTransferor.sol";
-import {CCIPLocalSimulator, IRouterClient, LinkToken, BurnMintERC677Helper} from "@chainlink/local/src/ccip/CCIPLocalSimulator.sol";
+import {
+    CCIPLocalSimulator,
+    IRouterClient,
+    LinkToken,
+    BurnMintERC677Helper
+} from "@chainlink/local/src/ccip/CCIPLocalSimulator.sol";
 
 contract TokenTransferorTest is Test {
     TokenTransferor public tokenTransferor;
@@ -14,20 +19,10 @@ contract TokenTransferorTest is Test {
 
     function setUp() public {
         ccipLocalSimulator = new CCIPLocalSimulator();
-        (
-            uint64 chainSelector_,
-            IRouterClient sourceRouter,
-            ,
-            ,
-            LinkToken linkToken,
-            BurnMintERC677Helper ccipBnM_,
+        (uint64 chainSelector_, IRouterClient sourceRouter,,, LinkToken linkToken, BurnMintERC677Helper ccipBnM_,) =
+            ccipLocalSimulator.configuration();
 
-        ) = ccipLocalSimulator.configuration();
-
-        tokenTransferor = new TokenTransferor(
-            address(sourceRouter),
-            address(linkToken)
-        );
+        tokenTransferor = new TokenTransferor(address(sourceRouter), address(linkToken));
 
         chainSelector = chainSelector_;
         ccipBnM = ccipBnM_;
@@ -40,34 +35,19 @@ contract TokenTransferorTest is Test {
 
         ccipBnM.drip(address(tokenTransferor));
 
-        ccipLocalSimulator.requestLinkFromFaucet(
-            address(tokenTransferor),
-            amountForFees
-        );
+        ccipLocalSimulator.requestLinkFromFaucet(address(tokenTransferor), amountForFees);
 
         tokenTransferor.allowlistDestinationChain(chainSelector, true);
 
         uint256 receiverBalanceBefore = ccipBnM.balanceOf(receiver);
-        uint256 tokenTransferorBalanceBefore = ccipBnM.balanceOf(
-            address(tokenTransferor)
-        );
+        uint256 tokenTransferorBalanceBefore = ccipBnM.balanceOf(address(tokenTransferor));
 
-        tokenTransferor.transferTokensPayLINK(
-            chainSelector,
-            receiver,
-            address(ccipBnM),
-            amountToSend
-        );
+        tokenTransferor.transferTokensPayLINK(chainSelector, receiver, address(ccipBnM), amountToSend);
 
         uint256 receiverBalanceAfter = ccipBnM.balanceOf(receiver);
-        uint256 tokenTransferorBalanceAfter = ccipBnM.balanceOf(
-            address(tokenTransferor)
-        );
+        uint256 tokenTransferorBalanceAfter = ccipBnM.balanceOf(address(tokenTransferor));
 
         assertEq(receiverBalanceAfter, receiverBalanceBefore + amountToSend);
-        assertEq(
-            tokenTransferorBalanceAfter,
-            tokenTransferorBalanceBefore - amountToSend
-        );
+        assertEq(tokenTransferorBalanceAfter, tokenTransferorBalanceBefore - amountToSend);
     }
 }